Marketing: factor-out Bullet component
